% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{add_displayed_classes}
\alias{add_displayed_classes}
\title{Add a Class to the 'Displayed' Ones}
\usage{
add_displayed_classes(x)
}
\arguments{
\item{x}{A character string, the name of a class.}
}
\description{
Add a class to those who are output using \code{display()} when
    they are autoprinted, i.e. returned as the value of the last expression
    in a Jupyter notebook cell.
}
