% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{add_http_handler}
\alias{add_http_handler}
\title{Add a handler for HTTP requests with slug given by `slug`.}
\usage{
add_http_handler(slug, handler)
}
\arguments{
\item{slug}{The first part (slug) of the URL to be handled}

\item{handler}{A function that handles GET requests with URLs that 
have `slug` as path component}
}
\description{
Add a handler for HTTP requests with slug given by `slug`.
}
