% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{alert}
\alias{alert}
\title{Create an alert in the frontend}
\usage{
alert(text)
}
\arguments{
\item{text}{A character string with the text that appears in the}
}
\description{
Creates Javascript code and sends it to the frontend that opens
    an alert box in the browser.
}
