% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-bytes.R
\name{Bytes}
\alias{Bytes}
\alias{BytesClass}
\title{Raw Bytes Traitlets}
\usage{
Bytes(...)
}
\arguments{
\item{...}{Arguments that are passed to the initialize method of 'BytesClass'}
}
\description{
A class and a constructor function to create (raw) bytes trait(let)s.
}
\section{Super class}{
\code{RKernel::Trait} -> \code{Bytes}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{A raw bytes vector}

\item{\code{optional}}{Logical, whether an initializing logical value must be provided}

\item{\code{coerce}}{Logical, whether 'as.raw()' is implicitely used when a value is
assigned to the trait}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Bytes-validator}{\code{BytesClass$validator()}}
\item \href{#method-Bytes-new}{\code{BytesClass$new()}}
\item \href{#method-Bytes-clone}{\code{BytesClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="get"><a href='../../RKernel/html/Trait.html#method-Trait-get'><code>RKernel::Trait$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="set"><a href='../../RKernel/html/Trait.html#method-Trait-set'><code>RKernel::Trait$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bytes-validator"></a>}}
\if{latex}{\out{\hypertarget{method-Bytes-validator}{}}}
\subsection{Method \code{validator()}}{
A validator method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BytesClass$validator(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A value to be checked for validity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bytes-new"></a>}}
\if{latex}{\out{\hypertarget{method-Bytes-new}{}}}
\subsection{Method \code{new()}}{
The initializing method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BytesClass$new(initial = raw(0), coerce = TRUE, optional = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{A value with which the traitlet is initialized}

\item{\code{coerce}}{Logical, used to initialize the 'coerce' field}

\item{\code{optional}}{Logical, used to initialize the 'optional' field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bytes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Bytes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BytesClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
