% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-options.R
\name{cell.options}
\alias{cell.options}
\title{Set options locally for the current jupyter notebook cell}
\usage{
cell.options(...)
}
\arguments{
\item{...}{Options, see \code{\link{options}}.}
}
\description{
Set options locally for the current jupyter notebook cell
}
