% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-options.R
\name{cell.par}
\alias{cell.par}
\title{Set graphics parameters locally for the current jupyter notebook cell}
\usage{
cell.par(...)
}
\arguments{
\item{...}{Graphics parameters, see \code{\link{par}}.}
}
\description{
Set graphics parameters locally for the current jupyter notebook cell
}
