% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-color.R
\name{ColorTrait}
\alias{ColorTrait}
\alias{ColorTraitClass}
\alias{Color}
\title{A Color String Trait}
\usage{
Color(...)
}
\arguments{
\item{...}{Arguments passed to the trait instance initializer}
}
\description{
A Color String Trait

A Color String Trait
}
\section{Super classes}{
\code{RKernel::Trait} -> \code{RKernel::Unicode} -> \code{Color}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{optional}}{Logical value, whether a length-zero value is allowed.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Color-validator}{\code{ColorTraitClass$validator()}}
\item \href{#method-Color-clone}{\code{ColorTraitClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="get"><a href='../../RKernel/html/Trait.html#method-Trait-get'><code>RKernel::Trait$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="set"><a href='../../RKernel/html/Trait.html#method-Trait-set'><code>RKernel::Trait$set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Unicode" data-id="initialize"><a href='../../RKernel/html/Unicode.html#method-Unicode-initialize'><code>RKernel::Unicode$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Color-validator"></a>}}
\if{latex}{\out{\hypertarget{method-Color-validator}{}}}
\subsection{Method \code{validator()}}{
Check the value assigned to the traitlet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorTraitClass$validator(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value assigned to the traitlet.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Color-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Color-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorTraitClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
