% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comms.R
\name{Comm}
\alias{Comm}
\alias{CommClass}
\title{Comms - connections between the kernel and the frontend}
\usage{
Comm(...)
}
\arguments{
\item{...}{Arguments passed to the inializer}
}
\description{
This R6 Class provides for bidirectional communication between the R Kernel and
   the Jupyter frontend, e.g. a Jupyter notebook
}
\details{
Objects of this class are used to communicate to the frontend via
\href{https://jupyter-client.readthedocs.io/en/latest/messaging.html#custom-messages}{custom messages}.
}
\section{Functions}{
\itemize{
\item \code{Comm()}: A constructor function for objects of class "CommClass"

}}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{A character string, the comm id}

\item{\code{target_name}}{A character string, the target}

\item{\code{handlers}}{A list of handler functions}

\item{\code{data}}{A list of data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Comm-new}{\code{CommClass$new()}}
\item \href{#method-Comm-open}{\code{CommClass$open()}}
\item \href{#method-Comm-send}{\code{CommClass$send()}}
\item \href{#method-Comm-close}{\code{CommClass$close()}}
\item \href{#method-Comm-clone}{\code{CommClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-new"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-new}{}}}
\subsection{Method \code{new()}}{
Initialize a 'Comm' object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommClass$new(target_name, id = uuid(), handlers = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_name}}{A string, the name of the target}

\item{\code{id}}{A string, the comm id}

\item{\code{handlers}}{A list of handler functions}

\item{\code{kernel}}{The relevant kernel}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-open"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-open}{}}}
\subsection{Method \code{open()}}{
Open a comm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommClass$open(data, metadata = emptyNamedList, buffers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A named list}

\item{\code{metadata}}{A named list}

\item{\code{buffers}}{A list of raw vectors or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-send"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-send}{}}}
\subsection{Method \code{send()}}{
Send data through a comm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommClass$send(data, metadata = emptyNamedList, buffers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A named list}

\item{\code{metadata}}{A named list}

\item{\code{buffers}}{A list of raw vectors or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-close"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-close}{}}}
\subsection{Method \code{close()}}{
Close a comm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommClass$close(
  data = emptyNamedList,
  metadata = emptyNamedList,
  buffers = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A named list}

\item{\code{metadata}}{A named list}

\item{\code{buffers}}{A list of raw vectors or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
