% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{CSS}
\alias{CSS}
\title{Send CSS code to the frontend}
\usage{
CSS(text, file)
}
\arguments{
\item{text}{A character string with CSS styling information}

\item{file}{Path of a file with CSS styling information}
}
\description{
Send CSS code in a character string or a text file to the
    frontend.
}
