% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTable.R
\name{dataTable}
\alias{dataTable}
\alias{dataTable.default}
\alias{dataTable.data.frame}
\alias{dataTableClass}
\alias{display_data.dataTable}
\title{HTML Tables with Interactive Controls}
\usage{
dataTable(x, ...)

\method{dataTable}{default}(x, ...)

\method{dataTable}{data.frame}(x, ...)

\method{display_data}{dataTable}(
  x,
  ...,
  metadata = emptyNamedList,
  id = attr(x, "id"),
  update = FALSE
)
}
\arguments{
\item{x}{A "dataTable" object}

\item{...}{Other arguments passed to the initialization method of
'dataTableClass' R6 objects}

\item{metadata}{A list of metadata strings}

\item{id}{An ID string}

\item{update}{A logical value; whether an existing display item will be updated}
}
\description{
Objects of class "dataTable" provide HTML tables with interactive
   controls powered by the DataTable Javascript library.
}
\section{Methods (by class)}{
\itemize{
\item \code{dataTable(default)}: Default method

\item \code{dataTable(data.frame)}: data.frame method

}}
\section{Methods (by generic)}{
\itemize{
\item \code{display_data(dataTable)}: dataTable method for display_data

}}
\section{Functions}{
\itemize{
\item \code{dataTable()}: A dataTable constructor

}}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{w}}{A container widget or NULL}

\item{\code{page}}{Number of the current page}

\item{\code{m}}{Width of the object divided by 'size'}

\item{\code{r}}{Remainder of the widht of the object divided by 'size'}

\item{\code{size}}{Number of columns in each group for horizontal paging}

\item{\code{iframe}}{An <iframe> container or NULL}

\item{\code{b_left}}{Button to scroll left}

\item{\code{b_right}}{Button to scroll right}

\item{\code{b_first}}{Button to scroll to the first group of columns}

\item{\code{b_last}}{Button to scroll to the last group of columns}

\item{\code{dt}}{HTML code for the visible table}

\item{\code{obj}}{The tabular object being dispayed}

\item{\code{label}}{A string label that shows the columns being displayed}

\item{\code{style}}{A string with CSS styling}

\item{\code{navigator}}{A container widget that contains the navigator buttons}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dataTable-new}{\code{dataTableClass$new()}}
\item \href{#method-dataTable-show_columns}{\code{dataTableClass$show_columns()}}
\item \href{#method-dataTable-page_left}{\code{dataTableClass$page_left()}}
\item \href{#method-dataTable-page_right}{\code{dataTableClass$page_right()}}
\item \href{#method-dataTable-page_first}{\code{dataTableClass$page_first()}}
\item \href{#method-dataTable-page_last}{\code{dataTableClass$page_last()}}
\item \href{#method-dataTable-draw}{\code{dataTableClass$draw()}}
\item \href{#method-dataTable-clone}{\code{dataTableClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-dataTable-new}{}}}
\subsection{Method \code{new()}}{
Initialize the DataTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableClass$new(obj, size = 50, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The object to be displayed}

\item{\code{size}}{An integer, the number of columns pre-formatted on each page.}

\item{\code{...}}{Other arguments, ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTable-show_columns"></a>}}
\if{latex}{\out{\hypertarget{method-dataTable-show_columns}{}}}
\subsection{Method \code{show_columns()}}{
Show which columns are displayed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableClass$show_columns()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTable-page_left"></a>}}
\if{latex}{\out{\hypertarget{method-dataTable-page_left}{}}}
\subsection{Method \code{page_left()}}{
Go one page to the left
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableClass$page_left()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTable-page_right"></a>}}
\if{latex}{\out{\hypertarget{method-dataTable-page_right}{}}}
\subsection{Method \code{page_right()}}{
Go one page to the right
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableClass$page_right()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTable-page_first"></a>}}
\if{latex}{\out{\hypertarget{method-dataTable-page_first}{}}}
\subsection{Method \code{page_first()}}{
Go to the first page (to the left)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableClass$page_first()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTable-page_last"></a>}}
\if{latex}{\out{\hypertarget{method-dataTable-page_last}{}}}
\subsection{Method \code{page_last()}}{
Go to the last page (to the right)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableClass$page_last()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTable-draw"></a>}}
\if{latex}{\out{\hypertarget{method-dataTable-draw}{}}}
\subsection{Method \code{draw()}}{
Draw the iframe with the data table contents
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableClass$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dataTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
