% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-datetime.R
\name{DateClass}
\alias{DateClass}
\alias{Date}
\alias{as.Date.DateClass}
\title{Date Traitlets}
\usage{
Date(...)

\method{as.Date}{DateClass}(x, ...)
}
\arguments{
\item{...}{Other arguments.}

\item{x}{A date traitlet.}
}
\description{
A class and constructor of date traitlets.
}
\section{Super class}{
\code{\link[RKernel:Trait]{RKernel::Trait}} -> \code{DateClass}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{A date.}

\item{\code{coerce}}{Logical value, whether assignments to the value field should
be coerced to the appropriate type.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DateClass-validator}{\code{DateClass$validator()}}
\item \href{#method-DateClass-new}{\code{DateClass$new()}}
\item \href{#method-DateClass-clone}{\code{DateClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="get"><a href='../../RKernel/html/Trait.html#method-Trait-get'><code>RKernel::Trait$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="set"><a href='../../RKernel/html/Trait.html#method-Trait-set'><code>RKernel::Trait$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DateClass-validator"></a>}}
\if{latex}{\out{\hypertarget{method-DateClass-validator}{}}}
\subsection{Method \code{validator()}}{
Check the value assigned to the traitlet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DateClass$validator(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value assigned to the traitlet.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DateClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-DateClass-new}{}}}
\subsection{Method \code{new()}}{
Initialize the traitlet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DateClass$new(
  initial = as.Date(integer(0)),
  year = integer(0),
  month = integer(0),
  day = integer(0),
  coerce = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{An optional Date object or date string}

\item{\code{year}}{An optional integer}

\item{\code{month}}{An optional integer}

\item{\code{day}}{An optional integer}

\item{\code{coerce}}{An optional logical value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DateClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DateClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DateClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
