% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\alias{[.dictionary}
\alias{[<-.dictionary}
\alias{print.dictionary}
\title{A data type analogous to Python dictionaries}
\usage{
dictionary(...)

\method{[}{dictionary}(x, i)

\method{[}{dictionary}(x, i) <- value

\method{print}{dictionary}(x, force = FALSE, ...)
}
\arguments{
\item{...}{Arbitrary objects. Should be tagged, yet currently name tags are
not yet checked for.}

\item{x}{A dictionary object}

\item{i}{A scalar integer or character string}

\item{value}{An arbitrary object}

\item{force}{A logical scalar, if TRUE, each element of the dictionary is printed,
if FALSE, just a brief summary is printed.}
}
\description{
Objects of class "dictionary" behave similar to dictionaries. They can
   contain any other kind of objects, but like with Python dictionaries, only scalar
   indices are allowed. Unlike with Python dictionaries, numeric indices can be used
   well as character indices.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: Get an element from a dictionary

\item \code{`[`(dictionary) <- value}: Set an element in a dictionary

\item \code{print(dictionary)}: Print a dictionary

}}
\section{Functions}{
\itemize{
\item \code{dictionary()}: A dictionary constructor

}}
