% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display an R Object}
\usage{
display(...)
}
\arguments{
\item{...}{Arguments passed to 'display_data' methods}
}
\description{
Sends a 'display_data' message to the frontend. 
 Allows users to create rich display of R objects.
}
