% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-env-browser.R
\name{envBrowser}
\alias{envBrowser}
\title{Environment Browser Widgets}
\usage{
envBrowser(
  pos = -1,
  name = NULL,
  envir,
  parent = NULL,
  all.names = FALSE,
  pattern = NULL,
  mode = "any"
)
}
\arguments{
\item{pos}{integer indicating the \code{\link[base]{search}} position, or -1
for the current environment.}

\item{name}{optional name indicating a position in the search path, see
\code{\link[base]{ls}}.}

\item{envir}{environment to use, see \code{\link[base]{ls}}.}

\item{parent}{an optional parent environment.}

\item{all.names}{logical; if true names starting with '.' are not omitted,
see \code{\link[base]{ls}}.}

\item{pattern}{an optional pattern of names to restrict browsing to, see
\code{\link[base]{ls}}.}

\item{mode}{an optional string indicating the mode of objects to which
browsing is restricted, see \code{\link[base]{ls}}.}
}
\description{
This function returns a widget that allows to browse within an environment
}
