% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{EventManager}
\alias{EventManager}
\alias{EventManagerClass}
\title{A Manager for Comms}
\usage{
EventManager(...)
}
\arguments{
\item{...}{Arguments passed to the inializer}
}
\description{
Objects of this class are used internally to manage events, they
    are not meant to be used by end-users.
}
\section{Functions}{
\itemize{
\item \code{EventManager()}: The constructor function, returns an Object of Class "EventManagerClass"

}}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventManager-new}{\code{EventManagerClass$new()}}
\item \href{#method-EventManager-send}{\code{EventManagerClass$send()}}
\item \href{#method-EventManager-on}{\code{EventManagerClass$on()}}
\item \href{#method-EventManager-activate}{\code{EventManagerClass$activate()}}
\item \href{#method-EventManager-suspend}{\code{EventManagerClass$suspend()}}
\item \href{#method-EventManager-clear}{\code{EventManagerClass$clear()}}
\item \href{#method-EventManager-resume}{\code{EventManagerClass$resume()}}
\item \href{#method-EventManager-has}{\code{EventManagerClass$has()}}
\item \href{#method-EventManager-clone}{\code{EventManagerClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventManager-new}{}}}
\subsection{Method \code{new()}}{
Initialize an event manager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventManagerClass$new(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A string, the type of event (e.g. "print")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventManager-send"></a>}}
\if{latex}{\out{\hypertarget{method-EventManager-send}{}}}
\subsection{Method \code{send()}}{
Send an event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventManagerClass$send(event, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{A string, the name of an event}

\item{\code{...}}{Other arguments, sent to the event handler(s)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventManager-on"></a>}}
\if{latex}{\out{\hypertarget{method-EventManager-on}{}}}
\subsection{Method \code{on()}}{
Install a handler for an event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventManagerClass$on(event, handler, remove = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{A string, the name of an event}

\item{\code{handler}}{A function}

\item{\code{remove}}{A logical value, whether the handler is to be removed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventManager-activate"></a>}}
\if{latex}{\out{\hypertarget{method-EventManager-activate}{}}}
\subsection{Method \code{activate()}}{
Activate handlers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventManagerClass$activate(event = NULL, all = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{A string, the name of an event, ignored if 'all' is TRUE.}

\item{\code{all}}{A logical value, if TRUE, all handlers that belong to 
the event type of the event manager are activated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventManager-suspend"></a>}}
\if{latex}{\out{\hypertarget{method-EventManager-suspend}{}}}
\subsection{Method \code{suspend()}}{
Suspend handlers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventManagerClass$suspend(event = NULL, all = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{A string, the name of an event, ignored if 'all' is TRUE.}

\item{\code{all}}{A logical value, if TRUE, all handlers that belong to 
the event type of the event manager are suspended.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventManager-clear"></a>}}
\if{latex}{\out{\hypertarget{method-EventManager-clear}{}}}
\subsection{Method \code{clear()}}{
Clear (i.e. remove) handlers for an event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventManagerClass$clear(event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{A string, the name of an event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventManager-resume"></a>}}
\if{latex}{\out{\hypertarget{method-EventManager-resume}{}}}
\subsection{Method \code{resume()}}{
Resume (i.e. reactivate) an event handler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventManagerClass$resume()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventManager-has"></a>}}
\if{latex}{\out{\hypertarget{method-EventManager-has}{}}}
\subsection{Method \code{has()}}{
Check whether the event manager has handlers for the
  given events
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventManagerClass$has(events)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{events}}{A character vector with names of events}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventManagerClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
