% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{fix_localhost}
\alias{fix_localhost}
\title{Adapt an localhost URL so that it works from behind Jupyterhub}
\usage{
fix_localhost(url)
}
\arguments{
\item{url}{A character string}
}
\description{
Adapt an localhost URL so that it works from behind Jupyterhub
}
