% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-interaction.R
\name{Fixed}
\alias{Fixed}
\title{Fixed Arguements in Interactive Widgets}
\usage{
Fixed(x)
}
\arguments{
\item{x}{An object.}
}
\description{
The function 'Fixed' returns ist argument 
  marked with a class attribute "Fixed", so that it
  is not made into a widget when passed to \code{\link{Interactive}}
}
