% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-numbers.R
\name{Float}
\alias{Float}
\alias{FloatClass}
\alias{as.integer.Float}
\alias{as.numeric.Float}
\title{Floating Point Number Traitlets}
\usage{
Float(...)

\method{as.integer}{Float}(x, ...)

\method{as.numeric}{Float}(x, ...)
}
\arguments{
\item{...}{Other arguments.}

\item{x}{A floating point traitlet.}
}
\description{
A class and a constructor function to create floating point vector trait(let)s.
}
\section{Super class}{
\code{\link[RKernel:Trait]{RKernel::Trait}} -> \code{Float}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{A numeric vector.}

\item{\code{optional}}{Logical value, whether a length-zero value is allowed.}

\item{\code{coerce}}{Logical value, whether assignments to the value field should
be coerced to the appropriate type.}

\item{\code{length}}{Integer number, the length the value should have.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Float-validator}{\code{FloatClass$validator()}}
\item \href{#method-Float-new}{\code{FloatClass$new()}}
\item \href{#method-Float-clone}{\code{FloatClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="get"><a href='../../RKernel/html/Trait.html#method-Trait-get'><code>RKernel::Trait$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="set"><a href='../../RKernel/html/Trait.html#method-Trait-set'><code>RKernel::Trait$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Float-validator"></a>}}
\if{latex}{\out{\hypertarget{method-Float-validator}{}}}
\subsection{Method \code{validator()}}{
Check the value assigned to the traitlet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatClass$validator(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value assigned to the traitlet.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Float-new"></a>}}
\if{latex}{\out{\hypertarget{method-Float-new}{}}}
\subsection{Method \code{new()}}{
Initialize the traitlet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatClass$new(
  initial = numeric(0),
  coerce = TRUE,
  optional = length(initial) == 0,
  length = 1L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{A numeric vector, the initial value for the traitlet.}

\item{\code{coerce}}{coerce Logical value, whether assignments to the value field should
be coerced to the appropriate type.}

\item{\code{optional}}{Logical value, whether a length-zero value is allowed.}

\item{\code{length}}{Integer number, the length the value should have.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Float-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Float-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
