% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completions.R
\name{get_completions}
\alias{get_completions}
\title{Provide completion for code given at point.}
\usage{
get_completions(code, cursor_pos)
}
\arguments{
\item{code}{A character string with code to be checked for
completions.}

\item{cursor_pos}{An integer, the current position of the cursor.}
}
\description{
Provide completion for code given at point.
}
