% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{has_http_handler}
\alias{has_http_handler}
\title{Check whether a handler exists for URLs that have `slug` as path component.}
\usage{
has_http_handler(slug)
}
\arguments{
\item{slug}{The first part (slug) of the URL to be handled}
}
\description{
Check whether a handler exists for URLs that have `slug` as path component.
}
