% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets.R
\name{HasTraits}
\alias{HasTraits}
\title{The Base Class of Objects with Traits}
\description{
Objects in class HasTraits have traits as components
   that are correctly initialized using delayed construction with
   the \code{\link{TraitInstance}} function.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{traits}}{A list of traits}

\item{\code{suspended}}{Logical value; whether notifying observers is suspended.}

\item{\code{observers}}{A list of observers, i.e. callback functions called by
the \code{notify} method.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HasTraits-new}{\code{HasTraits$new()}}
\item \href{#method-HasTraits-notify}{\code{HasTraits$notify()}}
\item \href{#method-HasTraits-observe}{\code{HasTraits$observe()}}
\item \href{#method-HasTraits-validate}{\code{HasTraits$validate()}}
\item \href{#method-HasTraits-clone}{\code{HasTraits$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HasTraits-new"></a>}}
\if{latex}{\out{\hypertarget{method-HasTraits-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HasTraits$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Initializsing values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HasTraits-notify"></a>}}
\if{latex}{\out{\hypertarget{method-HasTraits-notify}{}}}
\subsection{Method \code{notify()}}{
Notify observers about a trait being set to a value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HasTraits$notify(tn, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tn}}{A string, the name of the trait.}

\item{\code{value}}{The value to which the trait is set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HasTraits-observe"></a>}}
\if{latex}{\out{\hypertarget{method-HasTraits-observe}{}}}
\subsection{Method \code{observe()}}{
Install or remove an observer function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HasTraits$observe(tn, observer, remove = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tn}}{A string, the name of a trait.}

\item{\code{observer}}{A callback function, which should take three arguments,
(1) the trait name, (2) the object that has the trait, (3)}

\item{\code{remove}}{A logical value, indicates whether the observer is to be removed
or added}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HasTraits-validate"></a>}}
\if{latex}{\out{\hypertarget{method-HasTraits-validate}{}}}
\subsection{Method \code{validate()}}{
Install or remove the validator function of a trait.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HasTraits$validate(tn, validator, remove = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tn}}{A string, the name of a trait.}

\item{\code{validator}}{A callback function}

\item{\code{remove}}{A logical value, indicates whether the validator is to be removed
or added}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HasTraits-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HasTraits-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HasTraits$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
