% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{http_data}
\alias{http_data}
\title{Handle requests with slug "data".}
\usage{
http_data(path, query, ...)
}
\arguments{
\item{path}{The url, excluding the hostname and port, but including
the slug "data" and the name of an R object.}

\item{query}{The query string translated into a character vector.
The value of the "format" query
parameter should be one of "deparse", "str", "raw", "cat", "json", or 
"print". The query parameters "rows" and "cols" can be used to select
rows and columns.}

\item{...}{Any other arguments, ignored.}
}
\description{
These request can be used to get the value of an R object, usually a data frame.
}
\examples{
\donttest{
  browseURL(paste0(httpd_url(),"/data/iris?rows=1-5&format=json"))
}
}
