% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{http_echo}
\alias{http_echo}
\title{Handle requests with slug "echo".}
\usage{
http_echo(path, query, ...)
}
\arguments{
\item{path}{The url, excluding the hostname and port. Will be "/echo".}

\item{query}{The query string translated into a character vector.}

\item{...}{Any other arguments, ignored.}
}
\description{
A GET request from URL `http://localhost:XYZ/echo` will just
return the request.
}
