% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{http_eval}
\alias{http_eval}
\title{Handle requests with slug "eval".}
\usage{
http_eval(path, query, ...)
}
\arguments{
\item{path}{The url, excluding the hostname and port.}

\item{query}{The query string translated into a character vector.
The query argument named "expr" will be parsed and evaluated in a temporary environment.
The result of evaluated expression will be shown in a format
specified by the "format" query parameter. The value of the "format" query
parameter should be one of "deparse", "str", "raw", "cat", "json", or 
"print".}

\item{...}{Any other arguments, ignored.}
}
\description{
These request can be used to have R evaluate an expression
}
\examples{
\donttest{
 browseURL(paste0(httpd_url(),"/eval?expr=Sys.time()&format=json"))
} 
}
