% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{http_get}
\alias{http_get}
\title{Send a GET request to an URL and read the response.}
\usage{
http_get(x)
}
\arguments{
\item{x}{A character string with an URL.}
}
\description{
Send a GET request to an URL and read the response.
}
\examples{
\donttest{
 http_get(paste0(httpd_url(),"/data/iris?rows=1-5&format=json"))
}
}
