% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{http_is_proxied}
\alias{http_is_proxied}
\title{Check whether HTTP requests are proxied. This is the case if the kernel
is started from Jupyterhub and the Python package "jupyter_server_proxy"
is installed.}
\usage{
http_is_proxied()
}
\description{
If requests are proxied and the HTTP server of the R session listens at
port `<PORT>` then the appropriate URL is of the form 
`<JUPYTERHUB_SERVICE_PREFIX>/proxy/<PORT>` in addition to (or instead of)
`http://localhost:<PORT>`.
}
