% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{http_msg}
\alias{http_msg}
\title{Handle a generic message request sent via getting from an
url with the slug 'msg'.}
\usage{
http_msg(path, query, ...)
}
\arguments{
\item{path}{The url, excluding the hostname and port, but including
the slug "data" and the name of an R object.}

\item{query}{The query string translated into a character vector.}

\item{...}{Any other arguments, ignored.}
}
\description{
Handle a generic message request sent via getting from an
url with the slug 'msg'.
}
