% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{httpd_url}
\alias{httpd_url}
\title{Get the URL of the http server of the R session, including
the "/custom/" path component.}
\usage{
httpd_url()
}
\description{
Get the URL of the http server of the R session, including
the "/custom/" path component.
}
