% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{IFrame}
\alias{IFrame}
\title{Include HTML content using in an iframe}
\usage{
IFrame(url, width = "100\%", height = "70ex", class = NULL, srcdoc = FALSE)
}
\arguments{
\item{url}{A character string, the URL of the content to be included}

\item{width}{A character string that specifies the width of the iframe}

\item{height}{A character string that specifies the width of the iframe}

\item{class}{An optional character string with DOM classes to be assigned to the iframe.}

\item{srcdoc}{Logical, whether to use a 'src' (FALSE, the default)  or 'srcdoc' attribute.}
}
\description{
Display the contents of a webpage or other HTML content
  by including output using an [iframe](https://html.spec.whatwg.org/multipage/iframe-embed-object.html).
}
