% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap-helpers.R
\name{inspect_variables}
\alias{inspect_variables}
\title{Inspect variables for the debug adapter}
\usage{
inspect_variables(envir = parent.frame())
}
\arguments{
\item{envir}{The environment to be inspected}
}
\description{
Inspect variables for the debug adapter
}
