% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\alias{installspec}
\title{install the R Kernel}
\usage{
install()

installspec(user = TRUE, prefix = NULL, single_blas = FALSE)
}
\arguments{
\item{user}{Logical, whether to install the kernel in the user's home directory}

\item{prefix}{NULL or a character string with a path prefix}
}
\description{
install the R Kernel
}
\section{Functions}{
\itemize{
\item \code{installspec()}: Install the R Kernel spec

}}
