% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{Javascript}
\alias{Javascript}
\title{Send Javascript to the frontend}
\usage{
Javascript(text, file, as_tag = FALSE)
}
\arguments{
\item{text}{A character string with Javascript code}

\item{file}{Path of a file with Javascript code}

\item{as_tag}{Logical, whether to return a '<script>' tag}
}
\value{
An S3 object of class "display_data" with mime data of type "application/javascript"
}
\description{
Send Javascript code in a character string or a text file to the
    frontend.
}
