% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{LaTeXMath}
\alias{LaTeXMath}
\title{Send LaTeX math to the frontend}
\usage{
LaTeXMath(text)
}
\arguments{
\item{text}{A character string with LaTeX code for math}
}
\description{
Send LaTeX code for math in a character string to the
    frontend to be formatted by MathJax
}
