% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-vector.R
\name{List}
\alias{List}
\alias{ListClass}
\title{List Traitlets}
\usage{
List(...)
}
\arguments{
\item{...}{Arguments that are passed to the initialize method of 'ListClass'}
}
\description{
A class and a constructor function to create list trait(let)s.
}
\section{Super class}{
\code{\link[RKernel:Trait]{RKernel::Trait}} -> \code{List}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{A list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-List-validator}{\code{ListClass$validator()}}
\item \href{#method-List-clone}{\code{ListClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="get"><a href='../../RKernel/html/Trait.html#method-Trait-get'><code>RKernel::Trait$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="initialize"><a href='../../RKernel/html/Trait.html#method-Trait-initialize'><code>RKernel::Trait$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="set"><a href='../../RKernel/html/Trait.html#method-Trait-set'><code>RKernel::Trait$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-List-validator"></a>}}
\if{latex}{\out{\hypertarget{method-List-validator}{}}}
\subsection{Method \code{validator()}}{
A function that checks the validity of an assigned value, 
      i.e. whether the assigned value is a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListClass$validator(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A value to be assigned as the traitlet value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-List-clone"></a>}}
\if{latex}{\out{\hypertarget{method-List-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
