% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-browser.R
\name{ls_str}
\alias{ls_str}
\title{A HTML version of 'ls.str()'}
\usage{
ls_str(
  pos = -1,
  name = NULL,
  envir,
  all.names = FALSE,
  pattern = NULL,
  mode = "any"
)
}
\arguments{
\item{pos}{integer indicating \code{\link[base]{search}} path position, or -1 for the current environment.}

\item{name}{an optional name indicating search path position, see \code{\link[base]{ls}}.}

\item{envir}{the environment to look into}

\item{all.names}{logical value, if FALSE objects with names that start with a dot are ignored}

\item{pattern}{a character string, the pattern of the names of the objects to show}

\item{mode}{a character string, the mode of the objects to be shown}
}
\description{
This function is now just an alias for \code{\link{envBrowser}}.
}
\seealso{
\code{\link{ls.str}}
}
