% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-interaction.R
\name{mkWidget}
\alias{mkWidget}
\alias{mkWidget.integer}
\alias{mkWidget.numeric}
\alias{mkWidget.logical}
\alias{mkWidget.character}
\alias{mkWidget.Fixed}
\alias{mkWidget.ValueWidget}
\title{Controls for Interactive Widgets}
\usage{
mkWidget(x, ...)

\method{mkWidget}{integer}(x, description = NULL, ...)

\method{mkWidget}{numeric}(x, description = NULL, ...)

\method{mkWidget}{logical}(x, description = NULL, ...)

\method{mkWidget}{character}(x, description = NULL, ...)

\method{mkWidget}{Fixed}(x, ...)

\method{mkWidget}{ValueWidget}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{Other arguments, passed to more specific methods or ignored}

\item{description}{NULL or a character string that contains a description.}
}
\description{
A set of functions that can be used to create interactive
    widgets and to interact with widgets.
}
\details{
The function \code{mkWidget} is a generic function that creates a
    widget that allows to manipulate the arguments of a function that is
    called in an interactive widget. This generic function is called by the
    function \code{\link{mkWidgets}}. The function \code{Fixed} marks a value
    as fixed, so that \code{mkWidget} returns it as is.
}
