% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{Page}
\alias{Page}
\alias{Page.default}
\title{Display an object using the Jupyter notebook pager}
\usage{
Page(x, ...)

\method{Page}{default}(x, start = 1, ...)
}
\arguments{
\item{x}{An object to be displayed in the Notebook pager}

\item{...}{Other arguments, ignored or passed to specific methods.}

\item{start}{Integer, where to start the output.}
}
\description{
This function allows to display an \emph{R} object in the pager
    of a Jupyter notebook. Note that acts like \code{display()} when Jupyter
    Lab is used.
}
\section{Methods (by class)}{
\itemize{
\item \code{Page(default)}: S3 default method -- calls \code{display_data} and marks it as pager payload

}}
