% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets.R
\name{R6Class_}
\alias{R6Class_}
\title{R6 Objects That Are Not Locked}
\usage{
R6Class_(..., lock_objects = FALSE)
}
\arguments{
\item{...}{Arguments passed to the superclass constructor}

\item{lock_objects}{A logical value, indicates whether objects should be
locked. See \code{\link[R6]{R6Class}}.}
}
\description{
This function calls the R6 class constructor in such a way that
  new members can be added to the created objects.
}
