% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-R6instance.R
\name{R6Instance}
\alias{R6Instance}
\title{A Generic Constructor for R6 Object Traits}
\usage{
R6Instance(...)
}
\arguments{
\item{...}{Arguments passed to the trait instance initializer}
}
\description{
A Generic Constructor for R6 Object Traits
}
