% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic-codes.R
\name{register_magic_handler}
\alias{register_magic_handler}
\title{Register a handler for magics}
\usage{
register_magic_handler(magic, handler)
}
\arguments{
\item{magic}{A character string that selects a handler}

\item{handler}{A function that takes at least the argument 'code' and more '...' arguments.
The latter are constructed from the arguments of the percentage magic.}
}
\description{
Similar to '%%' magics in Jupyter/python it is possible to use such masics with this R kernel.
There are pre-defined magics for LaTeX math, CSS, Javascrippt, HTML, and iframes.
}
