% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{remove_paged_classes}
\alias{remove_paged_classes}
\title{Remove a Class to the 'Paged' Ones}
\usage{
remove_paged_classes(x)
}
\arguments{
\item{x}{A character string, the name of a class.}
}
\description{
Remove a class from those who are output using \code{Page()} when
    they are autoprinted, i.e. returned as the value of the last expression
    in a Jupyter notebook cell.
}
