% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{set_http_port}
\alias{set_http_port}
\title{Set the port the HTTP server listens to. For internal use only.}
\usage{
set_http_port(port)
}
\arguments{
\item{port}{An integer}
}
\description{
Set the port the HTTP server listens to. For internal use only.
}
