% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedHelp.R
\name{sharedHelpServer}
\alias{sharedHelpServer}
\title{R6 objects that run a help server}
\description{
An Object of class "sharedHelper" serves HTML help pages for one or several
  RKernel processes
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{port}}{Integer, the port number}

\item{\code{url}}{Character string, the base URL of the served help pages}

\item{\code{orig_httpd}}{A function, set to the original HTTP server function}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sharedHelpServer-new}{\code{sharedHelpServer$new()}}
\item \href{#method-sharedHelpServer-httpd}{\code{sharedHelpServer$httpd()}}
\item \href{#method-sharedHelpServer-run}{\code{sharedHelpServer$run()}}
\item \href{#method-sharedHelpServer-publish_port}{\code{sharedHelpServer$publish_port()}}
\item \href{#method-sharedHelpServer-log}{\code{sharedHelpServer$log()}}
\item \href{#method-sharedHelpServer-clone}{\code{sharedHelpServer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sharedHelpServer-new"></a>}}
\if{latex}{\out{\hypertarget{method-sharedHelpServer-new}{}}}
\subsection{Method \code{new()}}{
Intialize the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharedHelpServer$new(port = 0, prefix = "", use_proxy = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{Integer, a port number}

\item{\code{prefix}}{A character string, the URL prefix}

\item{\code{use_proxy}}{A logical value, whether the help server is supposed to be run
behind a jupyter proxy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sharedHelpServer-httpd"></a>}}
\if{latex}{\out{\hypertarget{method-sharedHelpServer-httpd}{}}}
\subsection{Method \code{httpd()}}{
The function that serves paths and queries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharedHelpServer$httpd(path, query, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{A character string, the path part of an URL}

\item{\code{query}}{An optional HTTP query string}

\item{\code{...}}{Any other arguments, passed on to the original 'httpd' function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sharedHelpServer-run"></a>}}
\if{latex}{\out{\hypertarget{method-sharedHelpServer-run}{}}}
\subsection{Method \code{run()}}{
The server loop
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharedHelpServer$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sharedHelpServer-publish_port"></a>}}
\if{latex}{\out{\hypertarget{method-sharedHelpServer-publish_port}{}}}
\subsection{Method \code{publish_port()}}{
Put a port number into a temporary file, for other processes to find
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharedHelpServer$publish_port(port)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{An integer, the port number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sharedHelpServer-log"></a>}}
\if{latex}{\out{\hypertarget{method-sharedHelpServer-log}{}}}
\subsection{Method \code{log()}}{
Put log text into a temporary file, for other processes to read
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharedHelpServer$log(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A character string to be added to the log file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sharedHelpServer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sharedHelpServer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharedHelpServer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
