% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str2iframe}
\alias{str2iframe}
\title{Create an HTML iframe tag that refers to some (usually HTML) code}
\usage{
str2iframe(
  code,
  resize = FALSE,
  width = "100\%",
  aspect_ratio = "16 / 10",
  height = character(0),
  class = "rkernel-iframe",
  style = "border-style:none",
  use_srcdoc = FALSE,
  ...
)
}
\arguments{
\item{code}{The code to be shown in the iframe}

\item{resize}{Logical; should the iframe be resizeable?}

\item{width}{The intended width of the iframe, a string or a number}

\item{aspect_ratio}{The intended aspect ratio of the iframe, a string}

\item{height}{The intended height, a string. Overrides the aspect ratio if given.}

\item{class}{The DOM class attribute the iframe, a string}

\item{style}{The CSS style attribte of the iframe, a string}

\item{use_srcdoc}{A logical value, whether the 'srcdoc' attribute should
be used to set the iframe content, rather than the 'src' attribute.}

\item{...}{Other arguments, ignored.}
}
\description{
Create an HTML iframe tag that refers to some (usually HTML) code
}
