% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-str_enum.R
\name{StrEnum}
\alias{StrEnum}
\title{An Enumerated String Constructor}
\usage{
StrEnum(...)
}
\arguments{
\item{...}{Arguments passed to the trait instance initializer}
}
\description{
An Enumerated String Constructor
}
