% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-str_enum.R
\name{StrEnumClass}
\alias{StrEnumClass}
\title{An Enumerated Strings Trait}
\description{
An Enumerated Strings Trait

An Enumerated Strings Trait
}
\section{Super classes}{
\code{\link[RKernel:Trait]{RKernel::Trait}} -> \code{\link[RKernel:Unicode]{RKernel::Unicode}} -> \code{StrEnum}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enum}}{a character vector}

\item{\code{optional}}{A logical value, whether value can be empty.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StrEnum-validator}{\code{StrEnumClass$validator()}}
\item \href{#method-StrEnum-new}{\code{StrEnumClass$new()}}
\item \href{#method-StrEnum-clone}{\code{StrEnumClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="get"><a href='../../RKernel/html/Trait.html#method-Trait-get'><code>RKernel::Trait$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="set"><a href='../../RKernel/html/Trait.html#method-Trait-set'><code>RKernel::Trait$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrEnum-validator"></a>}}
\if{latex}{\out{\hypertarget{method-StrEnum-validator}{}}}
\subsection{Method \code{validator()}}{
Check whether the assigned vector is one of the allowed enumerated strings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrEnumClass$validator(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A value to be assigned to the trait}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrEnum-new"></a>}}
\if{latex}{\out{\hypertarget{method-StrEnum-new}{}}}
\subsection{Method \code{new()}}{
Initialize the trait.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrEnumClass$new(enum, default = character(0), optional = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enum}}{A character vector of permitted enumerated strings.}

\item{\code{default}}{The default value}

\item{\code{optional}}{Logical can the value be empty?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrEnum-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StrEnum-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrEnumClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
