% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets.R
\name{Traitlets}
\alias{Traitlets}
\alias{TraitClass}
\alias{Trait}
\alias{TraitInstance}
\title{Traitlets}
\usage{
Trait(...)

TraitInstance(Class, ...)
}
\arguments{
\item{...}{Arguments passed to the inializer}

\item{Class}{An R6 Class that inherits from "TraitClass"}
}
\description{
The class \code{TraitClass} brings (some of) the functionality of the
    \href{https://traitlets.readthedocs.io/}{traitlets framework} on which
    the \href{https://ipywidgets.readthedocs.io}{ipywidgets framework} is
    based to \emph{R}.

The function \code{TraitInstance}  returns information needed by a
    \code{\link{HasTraits}} object to construct a \code{\link{TraitClass}}
    object.
}
\section{Functions}{
\itemize{
\item \code{Trait()}: A Baseline Trait Constructor

\item \code{TraitInstance()}: A "Delayed Constructor" for Traits, to be used by constructors of derived classes.

}}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{The value of the trait}

\item{\code{observers}}{A list of functions to be called as notification callbacks}

\item{\code{validators}}{A list of functions to check the validity of a}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Trait-set}{\code{TraitClass$set()}}
\item \href{#method-Trait-get}{\code{TraitClass$get()}}
\item \href{#method-Trait-new}{\code{TraitClass$new()}}
\item \href{#method-Trait-clone}{\code{TraitClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trait-set"></a>}}
\if{latex}{\out{\hypertarget{method-Trait-set}{}}}
\subsection{Method \code{set()}}{
Set the value of the trait
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraitClass$set(value, notify = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value to be set}

\item{\code{notify}}{Logical; whether to call notification callbacks}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trait-get"></a>}}
\if{latex}{\out{\hypertarget{method-Trait-get}{}}}
\subsection{Method \code{get()}}{
Get the trait value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraitClass$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trait-new"></a>}}
\if{latex}{\out{\hypertarget{method-Trait-new}{}}}
\subsection{Method \code{new()}}{
Initialize the trait, i.e. set an initial value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraitClass$new(initial)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{The initial value}

\item{\code{coerce}}{Logical; whether to coerce the initial value
to the approriate mode.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trait-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Trait-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraitClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
