% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-unicode.R
\name{Unicode}
\alias{Unicode}
\title{A Unicode String Trait Constructor}
\usage{
Unicode(...)
}
\arguments{
\item{...}{Arguments passed to the trait instance initializer}
}
\description{
A Unicode String Trait Constructor
}
