% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-unicode.R
\name{UnicodeClass}
\alias{UnicodeClass}
\alias{as.character.Unicode}
\title{A Unicode String Vector Trait}
\usage{
\method{as.character}{Unicode}(x, ...)
}
\arguments{
\item{x}{A Unicode traitlet}

\item{...}{Other arguments, ignored.}
}
\description{
A Unicode String Vector Trait

A Unicode String Vector Trait
}
\section{Functions}{
\itemize{
\item \code{as.character(Unicode)}: Coerce a unicode string trait to a character vector

}}
\section{Super class}{
\code{RKernel::Trait} -> \code{Unicode}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coerce}}{Logical value, whether values should be coerced to character strings.}

\item{\code{length}}{Length of the unicode character vector}

\item{\code{value}}{The value of the unicode character vector}

\item{\code{optional}}{A logical value, whethe valuer can be empty.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Unicode-validator}{\code{UnicodeClass$validator()}}
\item \href{#method-Unicode-new}{\code{UnicodeClass$new()}}
\item \href{#method-Unicode-clone}{\code{UnicodeClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="get"><a href='../../RKernel/html/Trait.html#method-Trait-get'><code>RKernel::Trait$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="set"><a href='../../RKernel/html/Trait.html#method-Trait-set'><code>RKernel::Trait$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unicode-validator"></a>}}
\if{latex}{\out{\hypertarget{method-Unicode-validator}{}}}
\subsection{Method \code{validator()}}{
A validator function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnicodeClass$validator(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value to be assigned}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unicode-new"></a>}}
\if{latex}{\out{\hypertarget{method-Unicode-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnicodeClass$new(
  initial = character(0),
  coerce = TRUE,
  optional = length(initial) == 0,
  length = 1L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{An initial value}

\item{\code{coerce}}{Logical value, whether values should be coerced to character stringes}

\item{\code{optional}}{Logical value, whether the value may be empty}

\item{\code{length}}{Integer, the intended length of the unicode vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unicode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Unicode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnicodeClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
