% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlets-vector.R
\name{Vector}
\alias{Vector}
\alias{VectorClass}
\title{Generic Vector Traitles}
\usage{
Vector(...)
}
\arguments{
\item{...}{Arguments that are passed to the initialize method of 'VectorClass'}
}
\description{
A class and a constructor function to create generic vector trait(let)s.
}
\section{Super class}{
\code{\link[RKernel:Trait]{RKernel::Trait}} -> \code{Vector}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{A list}

\item{\code{class}}{A character string, the common class of the vector elements}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Vector-validator}{\code{VectorClass$validator()}}
\item \href{#method-Vector-new}{\code{VectorClass$new()}}
\item \href{#method-Vector-clone}{\code{VectorClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="get"><a href='../../RKernel/html/Trait.html#method-Trait-get'><code>RKernel::Trait$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RKernel" data-topic="Trait" data-id="set"><a href='../../RKernel/html/Trait.html#method-Trait-set'><code>RKernel::Trait$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-validator"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-validator}{}}}
\subsection{Method \code{validator()}}{
A function that checks the validity of an assigned value, 
      i.e. whether the assigned value is a list with all elements of the same class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorClass$validator(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A value to be assigned as the traitlet value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-new"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorClass$new(class = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{String, optional common class}

\item{\code{...}}{Arguments passed to the superclass initializer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
