% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View.R
\name{View}
\alias{View}
\alias{View.default}
\alias{View.data.frame}
\title{Invoke a Data Viewer}
\usage{
View(x, title = deparse(substitute(x)), ...)

\method{View}{default}(x, title = deparse(substitute(x)), ...)

\method{View}{data.frame}(x, title = deparse(substitute(x)), ...)
}
\arguments{
\item{x}{An \R object which can be coerced to a data frame with non-zero numbers of rows and columns.}

\item{title}{A string used as title. Currently unused.}

\item{...}{Other arguments, ignored.}
}
\description{
This is a re-implementation of \code{\link[utils]{View}} that works within
Jupyter notebooks by leveraging the \href{https://ipywidgets.readthedocs.io}{Jupyter widgets} infrastructure
or by using the DataTable Javascript library. The latter is the case if the system option "View.backend"
is set to "dataTable" or if this option is not set. Otherwise a 'virtable_widget' is used.
}
