% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-virtual-table.R
\name{virtable_widget}
\alias{virtable_widget}
\alias{fmt_tab_section}
\alias{fmt_tab_section.default}
\alias{fmt_tab_section.tbl_df}
\title{Widgets to show tabular data}
\usage{
virtable_widget(
  x,
  pagesize = getOption("rkernel_view_size", c(10, 10)),
  continuous_update = TRUE
)

fmt_tab_section(x, i, j)

\method{fmt_tab_section}{default}(x, i, j)

\method{fmt_tab_section}{tbl_df}(x, i, j)
}
\arguments{
\item{x}{A tabular object, e.g. a data frame}

\item{pagesize}{Number of rows and columns that are shown by the widget.}

\item{continuous_update}{Logical, whether sliders should lead continuous updates.}

\item{i}{Integer values, indexing rows}

\item{j}{Integer values, indexing cols}
}
\description{
This is a widget that can be used to potentially large tabular data objects, such as data frames.
It is also a potential backend for the \code{\link{View}} function.
}
\section{Functions}{
\itemize{
\item \code{fmt_tab_section()}: Format a section of a tabular object to be used by \code{virtable_widget}.

\item \code{fmt_tab_section(default)}: Default method

\item \code{fmt_tab_section(tbl_df)}: Tibble method

}}
