% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R, R/display.R, R/help.R
\name{display_data.Widget}
\alias{display_data.Widget}
\alias{display_data}
\alias{display_data.default}
\alias{display_data.htmlwidget}
\alias{display_data.display_data}
\alias{display_data.update_display_data}
\alias{display_data.svg}
\alias{update.display_data}
\alias{display_data.data.frame}
\alias{display_data.data.set}
\alias{display_data.importer}
\alias{display_data.descriptions}
\alias{display_data.matrix}
\alias{display_data.html_elem}
\alias{display_data.shiny.tag}
\alias{display_data.shiny.tag.list}
\alias{display_data.iframe}
\alias{display_data.htmlTable}
\alias{display_data.tableHTML}
\alias{display_data.help_files_with_topic}
\alias{display_data.hsearch}
\title{Prepare an R Object for Being Displayed}
\usage{
\method{display_data}{Widget}(x, ..., metadata = emptyNamedList, id = uuid(), update = FALSE)

display_data(x, ...)

\method{display_data}{default}(
  x,
  ...,
  data,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{htmlwidget}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{display_data}(x, ...)

\method{display_data}{update_display_data}(x, ...)

\method{display_data}{svg}(x, ...)

\method{update}{display_data}(object, ...)

\method{display_data}{data.frame}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{data.set}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{importer}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{descriptions}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{matrix}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{html_elem}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{shiny.tag}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{shiny.tag.list}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{iframe}(
  x,
  ...,
  metadata = emptyNamedList,
  id = attr(x, "id"),
  update = FALSE
)

\method{display_data}{htmlTable}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{tableHTML}(
  x,
  ...,
  metadata = emptyNamedList,
  id = UUIDgenerate(),
  update = FALSE
)

\method{display_data}{help_files_with_topic}(
  x,
  ...,
  id = UUIDgenerate(),
  update = FALSE,
  embedded = FALSE,
  include_button = TRUE
)

\method{display_data}{hsearch}(x, ..., id = UUIDgenerate(), update = FALSE)
}
\arguments{
\item{x}{An object}

\item{...}{Optional arguments tagged by mime types with mime data}

\item{metadata}{A list with named elements, containing metadata}

\item{id}{An identifier string}

\item{update}{A logical value, whether a new display item should be created
or an existing one should be updated}

\item{data}{A list with named elements, containing mime data}

\item{object}{An object of class "display_data"}

\item{embedded}{A logical value, whether the help page should be shown embedded
in the Jupyter notebook}

\item{include_button}{A logical value, whether to include a button that
opens a new tab for the help page.}
}
\value{
An object of class "display_data"
}
\description{
A generic function that prepares R objects for display using \code{display()}
}
\section{Methods (by class)}{
\itemize{
\item \code{display_data(Widget)}: Method for jupyter widgets

\item \code{display_data(default)}: Default method

\item \code{display_data(htmlwidget)}: S3 method for html widgets

\item \code{display_data(display_data)}: S3 method for "display_data" objects

\item \code{display_data(update_display_data)}: S3 method for "update_display_data" objects

\item \code{display_data(svg)}: S3 method for "svg" objects

\item \code{display_data(data.frame)}: S3 method for class 'data.frame'

\item \code{display_data(data.set)}: S3 method for class 'data.set'

\item \code{display_data(importer)}: S3 method for class 'importer'

\item \code{display_data(descriptions)}: S3 method for class 'descriptions'

\item \code{display_data(matrix)}: S3 method for matrices

\item \code{display_data(html_elem)}: S3 method for "html_elem" objects (see \code{\link[memisc]{html}})

\item \code{display_data(shiny.tag)}: S3 methods for "shiny.tab" objects

\item \code{display_data(shiny.tag.list)}: S3 methods for "shiny.tab.list" objects

\item \code{display_data(iframe)}: S3 methods for "iframe" objects

\item \code{display_data(htmlTable)}: S3 methods for "htmlTable" objects

\item \code{display_data(tableHTML)}: S3 methods for "tableHTML" objects

\item \code{display_data(help_files_with_topic)}: S3 method for help pages

\item \code{display_data(hsearch)}: S3 method for results of 'help.search()'

}}
\section{Methods (by generic)}{
\itemize{
\item \code{update(display_data)}: "update" method for "display_data" objects

}}
