% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{http_get}
\alias{http_get}
\title{Send a http request to an URL and return the response.}
\usage{
http_get(x)
}
\arguments{
\item{x}{A character string containing the URL.}
}
\description{
Send a http request to an URL and return the response.
}
